backend/controller/requirements.txt
google-generativeai
groq
openai
pydantic
python-dotenv
